+++ packages/lib/src/typeChecking/typeCheck.ts
@@ -1,6 +1,5 @@
-import { failure } from "../utils"
+import { resolveTypeChecker } from "./resolveTypeChecker"
 import { AnyType, TypeToData } from "./schemas"
-import { resolveTypeChecker } from "./TypeChecker"
 import { TypeCheckError } from "./TypeCheckError"
 
 /**
@@ -12,9 +11,6 @@ import { TypeCheckError } from "./TypeCheckError"
  * @returns A TypeError if the check fails or null if no error.
  */
 export function typeCheck<T extends AnyType>(type: T, value: TypeToData<T>): TypeCheckError | null {
-  if (!type) {
-    throw failure("a type must be passed")
-  }
   const typeChecker = resolveTypeChecker(type)
 
   if (typeChecker.unchecked) {