+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.22.0
 
+- Added `tProp` syntax sugar for optional primitives with a default value.
+- Added `String`, `Number`, `Boolean`, `null`, `undefined` as aliases for primitive types.
 - Added `findParentPath`.
 
 ## 0.21.0